// e2e/enums/button.enums.ts
export enum Buttons {
  VIEW = "view",
  SIGN_IN = "sign in",
  CREATE_ACCOUNT = "create account",
  SEND_RESET_LINK = "send reset link",
  CREATE_PRODUCT = "create product",
  ADD_SELECTED_PRODUCT = "add selected product",
  VIEW_CARTED_PRODUCT = "view carted product",
  REMOVE_CARTED_PRODUCT = "remove carted product",
}