// e2e/enums/text-field.enums.ts
export enum TextField {
  EMAIL = "email",
  PASSWORD = "password",
  FULLNAME = "fullname",
  WORK_EMAIL = "work email",
  WORKEMAIL = "workemail",
  COMPANY_NAME = "companyname",
  CONFIRM_PASSWORD = "confirm-password",
  PRODUCT_NAME = "product name",
  PRODUCT_CODE = "product code",
  CATEGORY = "category",
  PRICE = "price",
  STOCK = "stock",
}